Read population multiparent

+++ R/kinship.R
@@ -56,10 +56,11 @@ kinship <- function(data_obj, geno_obj, type=c("overall"), n_cores=4,
   #        Determine if locus is numerical or character        #
   #                                                            #
   ##############################################################
-  num <- "1"
-  snp <- data_obj$geno_names$locus[[1]]
-  
-  if(grepl(num,snp)==TRUE) {
+  #num <- "1"
+  #snp <- data_obj$geno_names$locus[[1]]
+  num_locus <- is.numeric(data_obj$geno_names$locus[[1]])  
+
+  if(num_locus) {
     locus<-"num"
   } else {
     locus<-"char"
@@ -72,7 +73,7 @@ kinship <- function(data_obj, geno_obj, type=c("overall"), n_cores=4,
   #################################################################
 
   ##check to see if genoprobs have been calculated, if not calculate genotype probablities
-  if(!("calc_genoprob" %in% class(geno_obj)) & locus=="num" & pop=="MPP"){
+  if(!("calc_genoprob" %in% class(geno_obj)) && locus=="num" && pop=="MPP"){
     
     ### create map and genoprobs using geno file
     
@@ -84,7 +85,7 @@ kinship <- function(data_obj, geno_obj, type=c("overall"), n_cores=4,
     
     map$pos <- as.numeric(sapply(temp,"[",2)) #pulls second column and makes it a list
     
-    #data_obj$save_rds(map,"map.RData")
+    #data_obj$save_rds(map,"map.RDS")
     
     genoprobs <- probs_doqtl_to_qtl2(geno_obj,map=map,pos_column = "pos") #creates genotype probabilities from DOqtl...can only be used for DO genotype file
     
@@ -98,7 +99,7 @@ kinship <- function(data_obj, geno_obj, type=c("overall"), n_cores=4,
   #                                                               #
   #################################################################
   
-  if(!("calc_genoprob" %in% class(geno_obj)) & locus=="char" & pop=="MPP"){
+  if(!("calc_genoprob" %in% class(geno_obj)) && locus=="char" && pop=="MPP"){
     
     ### create map and genoprobs using geno file
     map <- data.frame(marker = data_obj$geno_names[[3]],chr = data_obj$geno_names[[3]], 
@@ -110,7 +111,7 @@ kinship <- function(data_obj, geno_obj, type=c("overall"), n_cores=4,
     
     map$pos <- as.list(data_obj$marker_location)
     
-    #data_obj$save_rds(map,"map.RData")
+    #data_obj$save_rds(map,"map.RDS")
     
     genoprobs <- probs_doqtl_to_qtl2(geno_obj,map=map,pos_column = "pos") #creates genotype probabilities from DOqtl...can only be used for DO genotype file
     
@@ -127,7 +128,7 @@ kinship <- function(data_obj, geno_obj, type=c("overall"), n_cores=4,
   #          Create probability and map file if RIL            #
   #                                                            #
   ##############################################################
-  if(!("calc_genoprob" %in% class(geno_obj)) & pop=="RIL"){
+  if(!("calc_genoprob" %in% class(geno_obj)) && pop=="RIL"){
     write_population(data_obj, geno_obj, filename = file.path(qtl_path, qtl_file), na = "")
     cross <- read.cross(format="csv", dir = qtl_path, qtl_file, genotypes=c(0,.5,1))
     unlink(file.path(qtl_path, qtl_file)) #delete the file
@@ -144,7 +145,7 @@ kinship <- function(data_obj, geno_obj, type=c("overall"), n_cores=4,
   #           Create probability and map file if 2PP           #
   #                                                            #
   ##############################################################
-  if(!("calc_genoprob" %in% class(geno_obj)) & pop=="2PP"){
+  if(!("calc_genoprob" %in% class(geno_obj)) && pop=="2PP"){
     write_population(data_obj, geno_obj, filename = file.path(qtl_path, qtl_file), 
     na = "")
     cross <- read.cross(format="csv", dir = qtl_path, qtl_file, genotypes=c(0,.5,1))
@@ -174,7 +175,7 @@ kinship <- function(data_obj, geno_obj, type=c("overall"), n_cores=4,
 
       map <- map_df_to_list(map,pos_column = "pos")
       
-      #data_obj$save_rds(map,"map.RData")
+      #data_obj$save_rds(map,"map.RDS")
       
       K <- calc_kinship(probs=genoprobs,type=type, cores=n_cores)
     }
